create and use interfaces for several existing classes
